### 2023-01-23 #5095 @Kevin-Mattheus-Moerman
@openjournals/dev can you see what is wrong here? :point_up: is it due to the `æ` symbol in the authors last name `Ræder`?
