### 2023-01-23 #5095 @danielskatz
@Kevin-Mattheus-Moerman - clicking on the error, I see "Problem with affiliations for Trygve Ræder, perhaps the affiliations index need quoting? (Theoj::Error)"
