### 2023-01-23 #5095 @danielskatz
looking at the raw file, I think the problem is the second line here
```
- name: Department of Physics, Technical University of Denmark, Denmark
index: 1
```
needs to be indented
