### 2023-01-24 #5095 @Kevin-Mattheus-Moerman
@trygvrad thanks for fixing that. I am trying to see if this work is in scope for JOSS. It is on the small side and the main function appears to be web app based visualisation. Although visualisation code can be in scope for JOSS, usually it also features analysis, or at least it is clear that the visualisation clearly impacts the science (or its communication) at hand. A good test is the "Would I cite this?" test. Essentially if your work is important for scientific research (a requirement for JOSS) then researchers would feel the need to cite it in their publications. This would be because the software is more than a simple utility tool (not in scope for JOSS) and directly influences the scientific outcomes. I can see that this work is relevant for "online communication", but does it go beyond that such that one would directly use this for research? If you feel this work is more than a utility tool, can you help sketch a clearer picture here (in a comment, and perhaps also in the paper) as to how this work is relevant/useful (directly) to scientific research. If it has used in studies (and perhaps cited) you can refer to such works too. Thanks.
