### 2023-01-25 #5095 @trygvrad
Hi @Kevin-Mattheus-Moerman,

Thank you for the quick response.
I see your concern, that supporting tools are less commonly published and cited than tools directly involved in data manipulation and analysis, and visualization appears to be a simple operation. In our field there has been a great need for supporting tools, due to the difficulty of relating the directions in real and reciprocal space involved in Bragg diffraction. However, there is, as you say, little tradition for attribution in this kind of work, and therefore less incentive to create, share, or publish.

The precise alignment of the instrument for any given sample is a complicated calculation
where the scattering vector (calculated from the X-ray energy, Bragg reflection, and lattice vectors in reciprocal space) determines a fixed axis around which the sample may rotate freely to describe a family of possible alignments with different imaging planes. The mathematical complexity forms the backbone of this application, so that the correct geometrical alignment can be presented in terms of angles of the goniometer. The parameter space of possible alignments is vast, given different possibilities of sample orientation, X-ray energy, Bragg reflection, and imaging plane. This tool allows for the simulation and exploration of this parameter space.

The need to access and visualize the sample-instrument alignment exists both during planning, execution and during analysis. I highlighted the need during planning in the manuscript, because that is where I have seen the greatest gains after the creation of this tool. When collecting data for the experiments in arXiv:2210.08366, one of seven days of experimental time at the X-ray free electron laser was lost due to a miscommunication of the angle of the sample mount. The angle was correct, but attributed to the wrong reference. This type of error has been a recurring theme in our field. The associated loss of scientific output is difficult to quantify, but it appears to be significant, and is addressed with the introduction of this tool.

For ongoing experiments, I have been informed that this tool typically occupies at least one monitor at the ESRF at all times (while DF-XRM experiments are being performed). I believe this results in significantly increased scientific output, although, again, this is difficult to quantify. This use touches a number of publications each year (See https://www.esrf.fr/home/UsersAndScience/Experiments/StructMaterials/id06---hard-x-ray-microscope/publications.html for previous publications from the DF-XRM microscope at ESRF. The microscope is currently being moved to a dedicated, and should operate at double throughput once it re-opens)

In our group we use this tool also to support analysis and teaching. However, I am not privy to how other groups operate, or other educators teach, so I cannot tell if others operate in the same way.

I hope you will find this within your scope, as a tool that "supports the functioning of research instruments or the execution of research experiments". I am confident that this tool has had, and continues to have significant scientific impact. However, as you no doubt are well aware, scientific impact does not always correlate with citations.
