### 2023-02-09 #5095 @Kevin-Mattheus-Moerman
@trygvrad thanks for the detailed response. I (we) fully agree scientific impact does not correlate with citations (and some scientists are a bit obsessed with citation metrics). I suppose what I was getting at is that something is clearly research software if it needed to reproduce the results (and if that is the case one would need to mention and cite it). Perhaps "planning tools" should be mentioned/cited more often like you hint at, and they do have scientific impact. I think you've given me enough detail in your comment to sketch the importance to research. The handling editor may still have concerns on scope but I'll now proceed to find a handling editor.
