### 2023-02-14 8d56bb6
fixed a bug when h (in Q=hkl) was negative
