### 2023-03-02 #5177 @taw10
Minor comments on the paper text:

- Line 24 "but is instead at an irrational angle of rotation": minor objection to "irrational".  I think it's meant to mean "not a low-index crystallographic direction", which I agree with.  However there's certainly no requirement for the direction to be irrational (in a mathematical sense).  Suggestion: replace with "..but instead can be at any orientation to the crystal", or similar.
- Figure 2: Should be made larger, because the table is too small to read when the main text is at a comfortable size.
- Line 79: Typo "diffracion" (should be "diffraction")
- Line 93: Pedantic nitpick: The facility where DF-XRM was used, as described in the referenced paper (Holstad et al 2022), is the Linac Coherent Light Source (LCLS).
- The syntax error noticed by the editorialbot should be fixed.
- DOIs should be added for some of the references.

Installation and functional checks to follow soon!
