### 2023-03-07 #5177 @taw10
Comments after the functional test:

- There is one step missing from the installation instructions for stand-alone installation, though it might be considered trivial: `pip install  -r requirements.txt`.  With this, the installation proceeds smoothly.
- Perhaps include instructions on how to run the Jupyter example, though this might also be seen as trivial ("load it into JupyterLab" or equivalent).
- Small fix needed in the Jupyter example (which worked, otherwise).  Under "Plot absorbance", line 4:
`xtl.Scatter.setup_scatter(type='xray', energy_kev=energy_kev)`
Change `type` to `scattering_type`.
- There are examples (which worked!), but no automated tests.  Since most of the code is concerned with visualisation, it might be quite difficult to test thoroughly.  But perhaps an automated end-to-end test would be appropriate.  e.g. generating a PDF from one of the example files, which exercises a lot of the code.
- Contribution guidelines are a little hard to find.  There is a sentence at the bottom of the page when running the application ("please report bugs to [address] or contribute directly at github."), but nothing in the README.md, and no CONTRIBUTING.md.
