### 2023-03-07 #5177 @marcocamma
**Functionality**

- maybe I missed it but I cannot find clear installation instructions. They are quite easy to guess for those who have similar installs in the past but they could be described in a separate INSTALL file or as part of the README.md. It should the different steps (clone repository, installing with `pip install -r requirements`. It could be mentioned the creation of the virtual environment.
- when I try to install on a new venv (created with `python3 -m venv my_env`, several errors are shown; I do not know (yet?) if this matters
