+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={11},
   pages={113703},
   year={2020},
+  doi={10.1063/5.0011961},
   publisher={AIP Publishing LLC}
 }
 
@@ -17,6 +18,7 @@
   number={1},
   pages={012007},
   year={2019},
+  doi={10.1088/1757-899X/580/1/012007},
   organization={IOP Publishing}
 }
 @article{simons2015dark,
@@ -27,6 +29,7 @@
   number={1},
   pages={1--6},
   year={2015},
+  doi={10.1038/ncomms7098},
   publisher={Nature Publishing Group}
 }
 
@@ -35,6 +38,7 @@
   title={Real-time imaging of acoustic waves in bulk materials with X-ray microscopy},
   author={Holstad, Theodor S and Dresselhaus-Marais, Leora E and R{\ae}der, Trygve Magnus and Kozioziemski, Bernard and van Driel, Tim and Seaberg, Matthew and Folsom, Eric and Eggert, Jon H and Knudsen, Erik Bergb{\"a}ck and Nielsen, Martin Meedom and others},
   journal={arXiv preprint arXiv:2211.01042},
+  doi={10.48550/arXiv.2211.01042},
   year={2022}
 }
 
@@ -59,6 +63,7 @@
   number={2},
   pages={762--769},
   year={2014},
+  doi={10.1107/S1600576714004506},
   publisher={International Union of Crystallography}
 }
 
@@ -71,6 +76,7 @@
   number={1},
   pages={9--17},
   year={2018},
+  doi={10.1038/s41566-017-0072-5},
   publisher={Nature Publishing Group}
 }
 
@@ -83,6 +89,7 @@
   volume={131},
   number={4},
   pages={040901},
+  doi={10.1063/5.0072399},
   year={2022},
   publisher={AIP Publishing LLC}
 }
@@ -102,10 +109,11 @@
   year      = 2007
 }
 
-@online{plotly,
+@misc{plotly,
   author = {Plotly},
   title = {Collaborative data science},
   publisher = {Plotly Technologies Inc.},
+  journal = {},	
   address = {Montreal, QC},
   year = {2015},
   url = {https://plot.ly}
