### 2023-03-08 6f44bff
added wheel to requirements to suppress error when installing fpdf
