### 2023-03-08 #5177 @trygvrad
@taw10  @marcocamma I added some install instructions that also show launching streamlit and/or jupyter

@marcocamma I was able to reproduce the error you found when installing the requirements. It did not impact performance but I added 'wheel' to the requirements, which will suppress the error.
