### 2023-03-08 d71eb61
update width of Figure 2
