### 2023-03-08 #5177 @marcocamma
@trygvrad , I really like the idea of the TOML file for the lenses ...
On my computer the white text on the light yellow background makes very difficult to see the "Upload a .lens file or select below" text (see screenshot below).
This happens only with the dark theme (unfortunately was the default so I had to spend time before realizing there was a text message!)

![2023 03 08_22h03_screenshot](https://user-images.githubusercontent.com/5097439/223852040-a0905a53-0d52-4eb3-8269-9ec1184e50ad.png)
