### 2023-03-08 #5177 @marcocamma
@trygvrad , nice work !
I understand the interest of a GUI, yours is also nice and polished but it would it make sense to highlight that the package can be used as "library" as shown in the jupyter notebook example !
I agree with @taw10 that some automatic testing would be good; if I understood well the jupyter example, there is not need to visualize to get values (reflection to use, lens distances, etc.) to use in automatic testing.
Precalculating few usecases and hardcoding the results in a test might be a way to make sure that future changes do not break the code. Generating a pdf can indeed be nice to exercise the "GUI" aspects.

I will keep trying things and exercise the different aspects a bit before closing the review
