+++ DF-XRM_vis_streamlit.py
@@ -13,6 +13,7 @@ from libs import absorption_plots
 from libs import draw_crystal_structures
 from libs import three_d_draw
 from libs import optics_geometry
+from libs import orientation_helpers
 def get_binary_file_downloader_html(bin_file, file_label='File'):
     '''
     provides a link to a file that can be downloaded
@@ -177,19 +178,13 @@ if uploaded_file is not None or crystal != 'Upload':
                            'd spacing [Å]': "{:.3f}",
                           }
                 ))
-
     #######################################################################
     ################################ get Q ################################
     #######################################################################
 
     st.sidebar.markdown('Scattering vector')
     hkl_str = st.sidebar.text_input('Q vector (h, k, l)', '('+li[0][0][1:-1]+')') #min, max, default
-    hkl_spli = hkl_str.strip('()').split(',')
-    h = int(hkl_spli[0])
-    k = int(hkl_spli[1])
-    l = int(hkl_spli[2])
-    params['hkl'] = np.array([h,k,l]) #'3, 3, 3'
-    Q = xtl.Cell.calculateQ(params['hkl'])[0]
+    params['hkl'], Q = orientation_helpers.Q_from_hkl_string(xtl, hkl_str)
 
     #######################################################################
     ####### the rotation of the crystal with respect to the sample ########
@@ -206,28 +201,13 @@ if uploaded_file is not None or crystal != 'Upload':
     front_hkl_str = st.sidebar.text_input('Exit surface ~(h,k,l) or ~[u,v,w,] (x-arrow in figure)', '') #min, max, default
     st.sidebar.write("For the exit surface, only the component orthogonal to sample 'up' is perserved")
 
-    if not up_hkl_str == '':
-        s = up_hkl_str.strip('[()]').split(',')
-        up_hkl = np.array([float(s[0]),float(s[1]),float(s[2])])
-        if '(' in up_hkl_str:
-            up_dir = xtl.Cell.calculateQ(up_hkl)[0]
-        else:
-            up_dir = xtl.Cell.calculateR(up_hkl)[0]
-    else:
-        up_dir = Q
+    up_dir = orientation_helpers.up_dir_from_string(xtl, up_hkl_str, Q)
+    if up_hkl_str == '':
         st.write(f'Sample up direction = Q')
+    
+    front_dir = orientation_helpers.front_dir_from_string(xtl, front_hkl_str, Q)
 
-    if not front_hkl_str == '':
-        s = front_hkl_str.strip('[()]').split(',')
-        front_hkl = np.array([float(s[0]),float(s[1]),float(s[2])])
-        if '(' in front_hkl_str:
-            front_dir = xtl.Cell.calculateQ(front_hkl)[0]
-        else:
-            front_dir = xtl.Cell.calculateR(front_hkl)[0]
-    else:
-        front_dir = np.cross(Q,np.array([0,0,1]))
-        if np.abs(np.sum(front_dir))<0.0001:
-            front_dir = np.cross(Q,np.array([0,1,0]))
+    if front_hkl_str == '':
         front_index = xtl.Cell.indexQ(front_dir)[0]
         front_index/=np.sqrt(np.sum(front_index**2))
         if front_index[0]%1 == 0 and front_index[1]%1 == 0 and front_index[2]%1 == 0:
@@ -238,24 +218,17 @@ if uploaded_file is not None or crystal != 'Upload':
 
     z_rot = -np.arctan2(up_dir[1],up_dir[0])*180/np.pi
[…]
