### 2023-03-09 82910a8
test for pytest
test_geom.py runs an automated test for the geometry
It also produces a pdf, which should be inspected manually and compared to previous commit
