+++ tests/test_geom.py
@@ -1,0 +1,299 @@
+import numpy as np
+import pandas as pd
+import sys, os
+import requests
+import Dans_Diffraction
+import plotly
+import toml
+import fpdf
+import plotly.io as pio
+import PIL.Image
+
+from libs import absorption_plots
+from libs import draw_crystal_structures
+from libs import three_d_draw
+from libs import optics_geometry
+from libs import orientation_helpers
+
+def get_cif_file(url):
+    if not os.path.exists('cif_files'):
+        os.makedirs('cif_files')
+    cif_file = 'cif_files/'+url.split('/')[-1]
+    if not os.path.isfile(cif_file):
+        r = requests.get(url)
+        open(cif_file , 'wb').write(r.content)
+    return cif_file
+
+
+
+
+def test_geometry():
+    # input
+    energy_kev = 17
+    sample_thickness = 150
+    params = {}
+    params['wavelength_in_um'] = 4.135667696*10**-15 *299792458 / energy_kev /1000*10**6 #7.2932e-5 #7.2932e-11 m = 17 keV
+    params['shape'] = np.array([5000,5000,sample_thickness])
+    url = 'http://www.crystallography.net/cod/1533978.cif'
+     
+    crystal = 'MnO3Y'
+    
+    hkl_str = '(1,0,1)'
+    up_hkl_str = '(0,0,1)'
+    front_hkl_str = '(0,1,0)'
+    
+    is_beam_norm = False# "Minimize distance beam travels through sample" 
+    beam_exit_hkl_str = '(0,1,0)' # only used if is_beam_norm == False
+    
+    lens_file_path = f'assets/lens_files/id06.lens'
+    mainx = "d_tot' = 5000" # sample-detector distance, "d_tot or d_tot' [mm]"
+
+    #############
+    # answers for tests
+    Q_answer = np.array([5490.7589723,  10219.38634624,  5900.16545805])
+    two_theta_answer = 8.664149507727018
+
+    z_rot_answer = 0
+    y_rot_answer = -90
+    x_rot_answer = 120
+    ##############
+    
+    
+    # setup 
+    cif_file = get_cif_file(url)
+    xtl = Dans_Diffraction.Crystal(cif_file)
+    xtl.generate_lattice()
+    xtl.Scatter.setup_scatter(scattering_type='xray', energy_kev=energy_kev)
+    
+    material_str = ''.join(xtl.cif['_chemical_formula_sum'].replace('2','$_2$').replace('3','$_3$').split(' '))
+    
+    cell_par = str(xtl.Cell)
+    cell_par = cell_par.replace('\n','').replace('\r','')
+    cell_par = cell_par.split('A = ')
+    # make absorption fig    
+    fig2d = absorption_plots.get_att_plot(xtl, energy_kev, sample_thickness)
+    fig2d.savefig('2dfig.png')
+    
+    # scattering table 
+    xtl.Scatter._scattering_max_twotheta = 45
+    reflections = xtl.Scatter.print_all_reflections().split('\n')
+    li = []
+    for reflection in reflections[3:-2]:
+        spli = reflection.split()
+        #lambda = 2dsin(theta)
+        d_spacing = params['wavelength_in_um']*10**4/(2*np.sin(np.pi/360*float(spli[-2])))
+        spl = reflection.split()
+        li.append([' '.join(spl[:-2]), float(spl[-2]), float(spl[-1]),float(f'{d_spacing:13.3f}')])
+        #st.write(reflection+f'  {d_spacing:13.3f}')
+
+    spl = reflections[2].split()
[…]
