### 2023-03-10 7b25825
test that compares images
