### 2023-03-10 85d2037
ensure reasonable colors in dark mode
Added a gray semi-transparent background image
It ensures that the background is not completely white or black
In this way the border between the figures and background is always clear

the old method (hard coding background color) only supported light theme
