### 2023-03-10 8514189
fix for plotly figure during test
The font used by plotly varies by platform: therefore hide the text when rendering the test case
