+++ tests/test_geom.py
@@ -35,16 +35,16 @@ def test_geometry():
     params['wavelength_in_um'] = 4.135667696*10**-15 *299792458 / energy_kev /1000*10**6 #7.2932e-5 #7.2932e-11 m = 17 keV
     params['shape'] = np.array([5000,5000,sample_thickness])
     url = 'http://www.crystallography.net/cod/1533978.cif'
-     
+
     crystal = 'MnO3Y'
-    
+
     hkl_str = '(1,0,1)'
     up_hkl_str = '(0,0,1)'
     front_hkl_str = '(0,1,0)'
-    
-    is_beam_norm = False# "Minimize distance beam travels through sample" 
+
+    is_beam_norm = False# "Minimize distance beam travels through sample"
     beam_exit_hkl_str = '(0,1,0)' # only used if is_beam_norm == False
-    
+
     lens_file_path = f'assets/lens_files/id06.lens'
     mainx = "d_tot' = 5000" # sample-detector distance, "d_tot or d_tot' [mm]"
 
@@ -56,29 +56,29 @@ def test_geometry():
     z_rot_answer = 0
     y_rot_answer = -90
     x_rot_answer = 120
-    
+
     im_2d_compare = PIL.Image.open("tests/2dfig.png")
     im_3d_compare = PIL.Image.open("tests/3dfig.png")
     im_optics_compare = PIL.Image.open("tests/fig_optics.png")
     ##############
-    
-    
-    # setup 
+
+
+    # setup
     cif_file = get_cif_file(url)
     xtl = Dans_Diffraction.Crystal(cif_file)
     xtl.generate_lattice()
     xtl.Scatter.setup_scatter(scattering_type='xray', energy_kev=energy_kev)
-    
+
     material_str = ''.join(xtl.cif['_chemical_formula_sum'].replace('2','$_2$').replace('3','$_3$').split(' '))
-    
+
     cell_par = str(xtl.Cell)
     cell_par = cell_par.replace('\n','').replace('\r','')
     cell_par = cell_par.split('A = ')
-    # make absorption fig    
+    # make absorption fig
     fig2d = absorption_plots.get_att_plot(xtl, energy_kev, sample_thickness)
     fig2d.savefig('2dfig.png')
-    
-    # scattering table 
+
+    # scattering table
     xtl.Scatter._scattering_max_twotheta = 45
     reflections = xtl.Scatter.print_all_reflections().split('\n')
     li = []
@@ -94,7 +94,7 @@ def test_geometry():
     df = pd.DataFrame(li, columns = [' '.join(spl[:-2]), '2𝜃', spl[-1]+' [%]', 'd-spacing [Å]'])
     pd.set_option('display.max_rows', 300)
     tab = str(df)
-    
+
     # orientation
     hkl_spli = hkl_str.strip('()').split(',')
     h = int(hkl_spli[0])
@@ -102,42 +102,42 @@ def test_geometry():
     l = int(hkl_spli[2])
     params['hkl'] = np.array([h,k,l]) #'3, 3, 3'
     Q = xtl.Cell.calculateQ(params['hkl'])[0]
-    
+
     up_dir = orientation_helpers.up_dir_from_string(xtl, up_hkl_str, Q)
     front_dir = orientation_helpers.front_dir_from_string(xtl, front_hkl_str, Q)
-    
+
     z_rot = -np.arctan2(up_dir[1],up_dir[0])*180/np.pi
     y_rot = -np.arctan2(up_dir[2],np.sqrt(up_dir[0]**2+up_dir[1]**2))*180/np.pi
 
-    
+
     x_rot = 0
 
     cr = orientation_helpers.crystal_rotation(z_rot, y_rot, x_rot)
     front_dir_r = cr.rotate(np.copy(front_dir))
-    
+
[…]
