### 2023-03-12 #5177 @trygvrad
Thank you @taw10  and @marcocamma  for all the comments,

I did a number of changes:

- I added a CONTRIMBUTE.md with information regarding contribution

- I corrected the errors in the manuscript

- I added an automated test. An automated test is a very good idea in this case as any changes to the master branch is automatically applied to the live streamlit page, which may at any time be used for an ongoing experiment. I therefore made quite an agressive test to act as a speed bump when making changes. The test compares both numerical values, the images, and makes a pdf that can be downloaded on the 'actions' page. A note on the test is included in CONTRIMBUTE.md.

- Thank you for alerting me to the fact that dark mode was not really functional. Apparently streamlit will default to dark mode if your browser is set to dark mode. I had changed the background color so that it was clear what was a figure and what was background when using the page in light mode, which had unintended consequences for dark mode. I changed my approach to this and the streamlit app now works in both light and dark mode.
