### 2023-03-27 #5177 @jgostick
Hi @trygvrad, can you take a look at the generated checklist items above?

Also, invesitigate the error reported for the plotly citation.  This seems to be rendering fine in the PDF, so I'm actually not sure what editorialbot is complaining about.
