### 2023-03-27 #5177 @taw10
Hi everyone, sorry for the delay - I went on vacation at an inconvenient time.  I've taken a look, and all my comments have been addressed.  I submitted a [tiny pull request](https://github.com/trygvrad/DF-XRM_viz/pull/5) to fix a typo (`pytnon` -> `python`), but it's obvious what the instructions intend.

Just a note: in my test, I installed using Python 3.11.2 on Fedora 37.  SciPy 1.8.1 wasn't available, but I made a tiny change to `requirements.txt` to use any later SciPy version, and it seemed to work fine with version 1.10.1.
