+++ DF-XRM_vis_streamlit.py
@@ -505,5 +505,5 @@ if uploaded_file is not None or crystal != 'Upload':
     st.markdown(get_binary_file_downloader_html('assets/blender_import_script.py', 'blender import script'), unsafe_allow_html=True)
 
 
-st.write("This is version 2.0.0. Souce code available at https://github.com/trygvrad/DF-XRM_viz under MIT lisence.")
+st.write("This is version 2.1.x. Souce code available at https://github.com/trygvrad/DF-XRM_viz under MIT lisence.")
 st.write("please report bugs to tmara@dtu.dk, or contribute directly at github.")
