### 2023-04-02 #5177 @trygvrad
**Additional Author Tasks After Review is Complete**

The version number is **v2.1.0**, (https://github.com/trygvrad/DF-XRM_viz/releases/tag/v2.1.0)
It is indexed at https://doi.org/10.5281/zenodo.7792903
