### 2023-04-02 #5177 @jgostick
I think this plotly error is ok, the PDF looks fine.
