### 2023-04-07 #5177 @trygvrad
@Kevin-Mattheus-Moerman  I updated the title at the zenodo archive and fixed the spelling.
My apologies for the errors.
Note that I did _not_ create a new release, so the spelling errors are still present in the version of the paper included in the zenodo archive. Let me know if you want me to create a new release (this will create a new DOI).
