### 2023-04-13 c12b67d
fix for width of plotly element on mobile
